Rename fbdx -> dx
