+++ server/index.js
@@ -21,7 +21,7 @@ import webpackHotMiddleware from 'webpack-hot-middleware'
 
 import DefaultServerConfig from './config'
 import webpackConfig from '../tools/webpack.client.dev'
-import { compileDev, startDev } from '../tools/fbdx'
+import { compileDev, startDev } from '../tools/dx'
 import { configureStore } from '../common/store'
 import reducer from '../common/createReducer'
 import createRoutes from '../common/routes/root'
