+++ tools/dx.js
@@ -2,7 +2,7 @@
 
 import chalk from 'chalk'
 import opn from 'opn'
-import {execSync} from 'child_process'
+import { execSync } from 'child_process'
 import path from 'path'
 
 function clearConsole () {
